Now linking MathJax using online version.
The MathJax library is huge (45MB for >3000 files) and I don't want that inside
my site or in this repo. An online version is available from their cloud
service, it loads just as fast as the local version and it's much lighter.

I, for one, welcome our new cloudly overlord.