Now I get it. It way the depency generation.
Dependencies where store in a set, structure that does not preserve order. Thus,
it's not the main source what appears in the first position, but any of the
source files. Since the .Six file path is generated from the main source file,
it is mandatory that it always appears in the first position