DOCUMENT_ROOT may not end with a slash
There is no guarantee that DOCUMENT_ROOT ends with a slash:
if that's not the case, the $sixtus variable may end up being
set to something like /var/wwwsixtus rather than the expected
/var/www/sixtus, and subsequent calls to require_once() will
fail because the path doesn't exist.

Always add a slash after DOCUMENT_ROOT to avoid the issue.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>